organising servlets in controller by subfolders
