+++ src/main/java/com/webdev/cheeper/controller/DatabaseHealthCheckServlet.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
 
-@WebServlet("/health")
+@WebServlet("/api/health")
 public class DatabaseHealthCheckServlet extends HttpServlet {
     
     @Override
