+++ src/main/java/com/webdev/cheeper/controller/onboarding/UsernameCheckServlet.java
@@ -1,4 +1,4 @@
-package com.webdev.cheeper.controller;
+package com.webdev.cheeper.controller.onboarding;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
